requirements.txt
python-telegram-bot==20.7
trilium-py==1.2.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
python-dateutil==2.8.2